refactor(ConnectionProvider): extract logic for event handlers into separate hook
