+++ client/src/features/connection/index.ts
@@ -1,4 +1,4 @@
 import ConnectionProvider from "./ConnectionProvider";
-import { ConnContext } from "./ConnContext";
+import { ConnContext } from "./ConnectionContext";
 
 export { ConnContext, ConnectionProvider };
